Enforcement for response content to be of type application/json, and added Swagger folder for swagger related code.
