+++ aspnet/src/SDP.TaskManagement.WebHost/Swagger/SwaggerServiceExtensions.cs
@@ -1,0 +1,43 @@
+﻿using Microsoft.OpenApi.Models;
+
+using System.Reflection;
+
+namespace SDP.TaskManagement.WebHost.Swagger;
+
+public static class SwaggerServiceExtensions
+{
+    /// <summary>
+    /// Custom handler for swagger configuration.
+    /// </summary>
+    public static IServiceCollection AddSwaggerConfiguration(this IServiceCollection services)
+    {
+        services.AddSwaggerGen(options =>
+        {
+            options.SwaggerDoc("v1", new OpenApiInfo
+            {
+                Version = "v1",
+                Title = "TaskManagementWebApp API",
+                Description = "API for our SDP project.",
+                Contact = new OpenApiContact
+                {
+                    Name = "Nicolas Limbeger Ferreira",
+                    Url = new Uri("https://github.com/NicolasLFerreira")
+                },
+                License = new OpenApiLicense
+                {
+                    Name = "MIT",
+                    Url = new Uri("https://opensource.org/licenses/MIT")
+                }
+            });
+
+            options.OperationFilter<RemovePlainTextContentTypesFilter>();
+
+            // XML comments
+            var xmlFile = $"{Assembly.GetExecutingAssembly().GetName().Name}.xml";
+            var xmlPath = Path.Combine(AppContext.BaseDirectory, xmlFile);
+            options.IncludeXmlComments(xmlPath);
+        });
+
+        return services;
+    }
+}
